Fix the stupids
