+++ app/src/main/AndroidManifest.xml
@@ -1,29 +1,29 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.whereuat.whereu">
+    package="xyz.whereuat.whereuat">
 
     <uses-permission android:name="android.permission.ACCESS_FINE_LOCATION" />
-    <uses-permission android:name="android.permission.INTERNET"/>
+    <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission android:name="android.permission.WAKE_LOCK" />
 
-    <permission android:name="com.whereuat.whereu.permission.C2D_MESSAGE"
+    <permission android:name="xyz.whereuat.whereuat.permission.C2D_MESSAGE"
         android:protectionLevel="signature" />
-    <uses-permission android:name="com.whereuat.whereu.permission.C2D_MESSAGE"/>
+    <uses-permission android:name="xyz.whereuat.whereuat.permission.C2D_MESSAGE" />
 
-    <uses-permission android:name="android.permission.READ_CONTACTS"/>
+    <uses-permission android:name="android.permission.READ_CONTACTS" />
 
     <application
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:supportsRtl="true"
-        android:theme="@style/AppTheme">
+        android:theme="@style/AppTheme" >
         <service android:name=".LocationProviderService" />
         <activity
-            android:name=".RouterActivity"
+            android:name="xyz.whereuat.whereuat.RouterActivity"
             android:label="@string/app_name"
             android:screenOrientation="portrait"
-            android:theme="@style/AppTheme">
+            android:theme="@style/AppTheme" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -50,7 +50,7 @@
             android:permission="com.google.android.c2dm.permission.SEND" >
             <intent-filter>
                 <action android:name="com.google.android.c2dm.intent.RECEIVE" />
-                <category android:name="com.whereuat.whereu" />
+                <category android:name="xyz.whereuat.whereuat" />
             </intent-filter>
         </receiver>
 
@@ -64,7 +64,7 @@
 
         <service
             android:name=".gcm.WuaInstanceIDListenerService"
-            android:exported="false">
+            android:exported="false" >
             <intent-filter>
                 <action android:name="com.google.android.gms.iid.InstanceID" />
             </intent-filter>
@@ -72,7 +72,7 @@
 
         <service
             android:name=".gcm.RegistrationIntentService"
-            android:exported="false">
+            android:exported="false" >
         </service>
     </application>
