--- app/src/main/java/com/whereuat/whereu/db/command/InsertCommand.java
@@ -1,24 +1,0 @@
-package com.whereuat.whereu.db.command;
-
-import android.content.ContentValues;
-import android.content.Context;
-
-/**
- * Created by kangp3 on 4/5/16.
- */
-public class InsertCommand extends DbCommand {
-    private String mNullColumn;
-    private ContentValues mValues;
-
-    public InsertCommand(Context context, String table,
-                         String null_column, ContentValues values) {
-        super(context, table);
-        mNullColumn = null_column;
-        mValues = values;
-    }
-
-    @Override
-    public Long execute() {
-        return mDb.insert(mTable, mNullColumn, mValues);
-    }
-}
